requirements.txt
absl-py==0.15.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
arrow==1.2.3
asciitree==0.3.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blessed==1.20.0
blinker==1.4
Bottleneck==1.3.2
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.0.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
croniter==1.3.8
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dateutils==0.6.12
dbus-python==1.2.18
debugpy==1.6.6
decorator==4.4.2
deepdiff==6.3.0
defer==1.0.6
defusedxml==0.7.1
dicomsdl==0.109.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
docstring-parser==0.15
duplicity==0.8.21
efficientnet-pytorch==0.7.1
einops==0.6.1
email-validator==1.3.1
entrypoints==0.4
et-xmlfile==1.0.1
fastapi==0.88.0
fasteners==0.14.1
filelock==3.6.0
flake8==4.0.1
flatbuffers===1.12.1-git20200711.33e2d80-dfsg1-0.6
fonttools==4.29.1
frozenlist==1.3.3
fs==2.4.12
fsspec==2023.3.0
future==0.18.2
gast==0.5.2
gdown==4.7.1
glob2==0.7
google-auth==1.5.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.30.2
h11==0.14.0
h5py==3.6.0
h5py.-debian-h5py-serial==3.6.0
html5lib==1.1
httpcore==0.16.3
httplib2==0.20.2
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.12.0
hydra-core==1.3.2
idna==3.3
imagecodecs==2023.3.16
imageio==2.25.0
imgaug==0.4.0
importlib-metadata==4.6.4
importlib-resources==5.12.0
iniconfig==1.1.1
inquirer==3.1.3
install==1.3.5
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==8.0.4
iterative-stratification==0.1.7
itsdangerous==2.1.2
jdcal==1.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.2.0
jsonargparse==4.20.1
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.5
kaggle==1.5.12
keras==2.11.0
keyring==23.5.0
kiwisolver==1.3.2
kornia==0.6.10
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightning==2.0.0
lightning-cloud==0.5.32
lightning-utilities==0.8.0
lion-pytorch==0.0.6
llvmlite==0.38.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==2.2.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mdurl==0.1.2
monai==1.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
multidict==6.0.4
munch==2.5.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==3.0
nose==1.3.7
notebook==6.4.8
numba==0.55.1
numcodecs==0.11.0
numexpr==2.8.1
numpy==1.21.5
nvidia-dali-cuda110==1.16.0
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
odfpy==1.4.2
olefile==0.46
omegaconf==2.3.0
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
openpyxl==3.0.9
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.8.9
packaging==21.3
pandas==1.2.5
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
pluggy==0.13.0
ply==3.11
pretrainedmodels==0.7.4
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
pydantic==1.10.7
pydicom==2.3.1
pyflakes==2.4.0
Pygments==2.14.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pylibjpeg==1.4.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.2.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-editor==1.0.4
python-gdcm==3.0.20
python-multipart==0.0.6
python-slugify==7.0.0
pythran==0.10.0
pytorch-lightning==2.0.0
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
qudida==0.0.4
Random-Word==1.0.11
readchar==4.0.5
regex==2022.10.31
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==13.3.3
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
segmentation-models-pytorch==0.3.2
Send2Trash==1.8.1b0
shapely==2.0.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0.post2
sniffio==1.3.0
soupsieve==2.3.1
starlette==0.22.0
starsessions==1.3.0
sympy==1.9
systemd-python==234
tables==3.7.0
tensorboard==2.11.0
tensorboardX==2.6
tensorflow-estimator==2.11.0
tensorflow-gpu==2.11.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.3.21
timm==0.6.12
tokenizers==0.13.2
toml==0.10.2
torch==2.0.1
torchattacks==3.3.0
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.1
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.0
ttach==0.0.3
typeshed-client==2.2.0
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.7.0
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.21.1
uvloop==0.17.0
virtualenv==20.13.0+ds
volumentations-3D==1.0.4
wadllib==1.3.6
warmup-scheduler==0.3
watchfiles==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.0.2
widgetsnbextension==4.0.5
wrapt==1.13.3
xdg==5
xkit==0.0.0
xlwt==1.3.0
yarl==1.8.2
zarr==2.14.2
zipp==1.0.0
